Docstrings

HAdaptiveIntegration.CUBE_BE115 — Constant

Berntsen-Espelid with 115 nodes.

HAdaptiveIntegration.CUBE_BE65 — Constant

Berntsen-Espelid with 65 nodes.

HAdaptiveIntegration.CUBE_GM33 — Constant

Genz-Malik with 33 nodes.

HAdaptiveIntegration.SEGMENT_GK15 — Constant

Gauss-Kronrod with 15 nodes.

HAdaptiveIntegration.SEGMENT_GK31 — Constant

Gauss-Kronrod with 31 nodes.

HAdaptiveIntegration.SEGMENT_GK7 — Constant

Gauss-Kronrod with 7 nodes.

HAdaptiveIntegration.SQUARE_CH21 — Constant

Cools-Haegemans with 21 nodes.

HAdaptiveIntegration.SQUARE_CH25 — Constant

Cools-Haegemans with 25 nodes.

HAdaptiveIntegration.SQUARE_GM17 — Constant

Genz-Malik with 17 nodes.

HAdaptiveIntegration.TETRAHEDRON_GM35 — Constant

Grundmann-Möller with 35 nodes.

HAdaptiveIntegration.TRIANGLE_GM19 — Constant

Grundmann-Möller with 19 nodes.

HAdaptiveIntegration.TRIANGLE_RL19 — Constant

Radon-Laurie with 19 nodes.

HAdaptiveIntegration.AbstractDomain — Type

abstract type AbstractDomain{D,T}

Abstract type for integration domains' in D dimensions with element type T.

Type Parameters:

D: dimension of the domain.

T: element type of the domain.

Mandatory methods:

map_from_reference

abs_det_jac

Useful (but non-mandatory) methods:

reference_domain

map_to_reference

HAdaptiveIntegration.AbstractRule — Type

abstract type AbstractRule{DOM<:AbstractDomain}

Abstract type for a cubature rule on a domain DOM.

Type Parameters:

DOM: [reference_domain](@ref)(DOM) gives the reference domain on which the embedded cubature is assume to be set.

Mandatory methods:

embedded_cubature

Useful (but non-mandatory) methods:

orders

HAdaptiveIntegration.Cuboid — Type

Cuboid{T} = Orthotope{3,T}

Alias for a 3-dimensional cuboid of value type T.

HAdaptiveIntegration.EmbeddedCubature — Type

struct EmbeddedCubature{D,T}

An embedded cubature rule consisting of a high order cubature rule with H nodes and a low order cubature rule with L nodes. Note that the low order cubature uses nodes[1:L] as its nodes. The cubature nodes and weights are assumed to be for the reference domain.

Fields:

nodes::Vector{SVector{D,T}}: the cubature nodes.

weights_high::Vector{T}: the cubature weights for the high order cubature.

weights_low::Vector{T}: the cubature weights for the low order cubature.

Invariants (check at construction):

length(nodes) == length(weights_high)

length(weights_high) ≥ length(weights_low)

HAdaptiveIntegration.EmbeddedCubature — Method

(ec::EmbeddedCubature{D,T})(
    fct, domain::Domain{D,T}, norm=x -> LinearAlgebra.norm(x, Inf)
) where {D,T}

Return I_high and norm(I_high - I_low) where I_high and I_low are the result of the high order cubature and the low order cubature on domain. The function fct must take a SVector{D,T} to a return type K, and K must support the multiplication by a scalar and the addition. Note that there is no check, beyond compatibility of dimension and type, that the embedded cubature is for the right domain.

HAdaptiveIntegration.GenzMalik — Type

struct GenzMalik{D} <: AbstractRule{Orthotope{D}}

Embedded cubature rule for a D-orthotope of high order 7 and low order 5.

Type Parameters:

D: The dimension of the orthotope.

HAdaptiveIntegration.GrundmannMoeller — Type

struct GrundmannMoeller{D} <: AbstractRule{Simplex{D}}

Embedded cubature rule for a D-simplex.

Type Parameters:

D: The dimension of the simplex.

Fields:

order_high::Int: the high order of the rule.

order_low::Int: the low order of the rule.

Invariants (check at construction):

order_high and order_low must be odd.

must have order_high > order_low ≥ 1.

HAdaptiveIntegration.Orthotope — Type

struct Orthotope{D,T} <: AbstractDomain{D,T}

Axes-aligned Orthotope in D dimensions, with element type T, given by two points low_corner and high_corner. Note that, we must have low_corner .≤ high_corner.

Fields:

low_corner::SVector{D,T}: the low corner.

high_corner::SVector{D,T}: the high corner.

Invariants (not check at construction):

low_corner .≤ high_corner

HAdaptiveIntegration.RadonLaurie — Type

struct RadonLaurie <: AbstractRule{Simplex{2}}

Embedded cubature rule for a 2-simplex of high order 8 and low order 5.

HAdaptiveIntegration.Rectangle — Type

Rectangle{T} = Orthotope{2,T}

Alias for a 2-dimensional rectangle of element type T.

HAdaptiveIntegration.Segment — Type

Segment{T} = Orthotope{1,T}

Alias for a 1-dimensional segment of element type T.

HAdaptiveIntegration.Simplex — Type

struct Simplex{D,N,T} <: AbstractDomain{D,T}

A simplex in D dimensions with N=D+1 vertices of element type T.

Fields:

vertices::SVector{N,SVector{D,T}}: vertices of the simplex.

Invariants (not check at construction):

N = D+1

HAdaptiveIntegration.TabulatedEmbeddedCubature — Type

@kwdef struct TabulatedEmbeddedCubature{DOM<:AbstractDomain} <: AbstractRule{DOM}

An embedded cubature rule consisting of a high order cubature rule and a low order cubature rule. Note that the low order cubature uses nodes[1:L] as its nodes where L is the length of the weights_low. The cubature nodes and weights are assumed to be for the reference domain (use the reference_domain function to get the reference domain).

Fields:

description::String: description of the embedded cubature.

reference::String: where the values are found.

precision::Int: number of significant digits on the node and weight values, 10^-precision give the relative precision of the values.

nodes::Vector{Vector{String}}: the cubature nodes.

weights_high::Vector{String}: the cubature weights for the high order cubature.

order_high::Int: order of the high order cubature.

weights_low::Vector{String}: the cubature weights for the low order cubature.

order_low::Int: order of the low order cubature.

Invariants (check at construction):

length(nodes) == length(weights_high)

length(weights_high) ≥ length(weights_low)

order_high ≥ order_low ≥ 0

precision ≥ 0

HAdaptiveIntegration.Tetrahedron — Type

Tetrahedron{T} = Simplex{3,4,T}

Alias for a 3-dimensional tetrahedron with 4 vertices of element type T.

HAdaptiveIntegration.Triangle — Type

Triangle{T} = Simplex{2,3,T}

Alias for a 2-dimensional triangle with 3 vertices of value type T.

HAdaptiveIntegration.abs_det_jac — Method

abs_det_jac(domain::DOM) where {DOM<:AbstractDomain}

Return the absolute value of the Jacobian's determinant of the map from the reference domain to the physical domain domain.

HAdaptiveIntegration.allocate_buffer — Method

allocate_buffer(fct, domain, ec=default_embedded_cubature(domain))

Allocate and return a buffer that can be passed to the integrate function to improve performance by reducing memory allocations when integrate is called multiple times.

HAdaptiveIntegration.combinations — Method

combinations(n::Int, k::Int)

Helper function to generate all combinations of k elements from 1:n, similar to calling combinations(1:n, k) from Combinatorics.jl.

HAdaptiveIntegration.cuboid — Method

cuboid(low_corner, high_corner)
cuboid(T::DataType, low_corner, high_corner)

Return an axes-aligned cuboid, with element type T, given by two 3d-points low_corner and high_corner. Note that, we must have low_corner .≤ high_corner.

HAdaptiveIntegration.default_embedded_cubature — Method

default_embedded_cubature(domain::DOM) where {DOM<:AbstractDomain}

Return a default embedded cubature for the domains:

dimension 1:

segment: SEGMENT_GK15

dimension 2:

triangle: TRIANGLE_RL19

rectangle: SQUARE_CH25

dimension 3:

tetrahedron: TETRAHEDRON_GM35

cuboid: CUBE_BE65

dimension d:

simplex: GrundmannMoeller{d}(7, 5)`

orthotope: GenzMalik{d}()

HAdaptiveIntegration.default_subdivision — Method

default_subdivision(domain::DOM) where {DOM<:AbstractDomain}

Return the default algorithm to subdivide domain.

dimension 1:

segment: subdivide_segment2

dimension 2:

triangle: subdivide_triangle4

rectangle: subdivide_rectangle4

dimension 3:

tetrahedron: subdivide_tetrahedron8

cuboid: subdivide_cuboid8

dimension d:

simplex: subdivide_simplex

orthotope: subdivide_orthotope

HAdaptiveIntegration.dimension — Method

dimension(::Type{DOM}) where {DOM<:AbstractDomain{D}}

Return the dimension D of the given domain DOM.

HAdaptiveIntegration.embedded_cubature — Method

embedded_cubature(ar::AbstractRule)
embedded_cubature(T::DataType, ar::AbstractRule)

embedded_cubature(nodes, weights_high, weights_low)
embedded_cubature(T::DataType, nodes, weights_high, weights_low)

Construct the embedded cubature with element type T from an AbstractRule (TabulatedEmbeddedCubature or GrundmannMoeller), or from a vector of nodes and two vectors of weights for the high and low order cubature, with element type T.

HAdaptiveIntegration.integral_monomial_orthotope — Method

integral_monomial_orthotope(dim::Int, k_max::Int)

Return the values of monomial's integral over the reference orthotope. It return a Vector{ Vector{ Pair{ Ntuple{dim,Int}, Rational{Int} } } }. The outer vector is index by the total degree + 1, for the total degree form 0 to tot_deg_max. The inner vector contain Pair{ Ntuple{dim,Int}, Rational{Int} } where the Ntuple{D,Int} is the multi-index of the monomial and Rational{Int} is the value of the integral.

HAdaptiveIntegration.integral_monomial_simplex — Method

integral_monomial_simplex(dim::Int, k_max::Int)

Return the values of monomial's integral over the reference simplex. It return a Vector{ Vector{ Pair{ Ntuple{dim,Int}, Rational{Int} } } }. The outer vector is index by the total degree + 1, for the total degree form 0 to tot_deg_max. The inner vector contain Pair{ Ntuple{dim,Int}, Rational{Int} } where the Ntuple{D,Int} is the multi-index of the monomial and Rational{Int} is the value of the integral.

HAdaptiveIntegration.integrate — Method

integrate(
    fct,
    domain::AbstractDomain{D,T};
    embedded_cubature::EmbeddedCubature{D,T}=default_embedded_cubature(domain),
    subdiv_algo=default_subdivision(domain),
    buffer=nothing,
    norm=x -> LinearAlgebra.norm(x, Inf),
    atol=zero(T),
    rtol=(atol > zero(T)) ? zero(T) : sqrt(eps(T)),
    maxsubdiv=8192 * 2^D,
) where {D,T}

Return I and E where I is the integral of the function fct over domain and E is an error estimate.

Arguments

fct: a function that must take a SVector{D,T} to a return type K, with K must support the multiplication by a scalar of type T and the addition.

domain::AbstractDomain{D,T}: the integration domain. Currently, we support segment, triangle, rectangle, tetrahedron, cuboid, and d-dimensional simplex.

Optional arguments

embedded_cubature::EmbeddedCubature{D,T}=default_embedded_cubature(domain): the embedded cubature, each supported domain has a default_embedded_cubature.

subdiv_algo=default_subdivision(domain): the subdivision algorithm, each domain has a default_subdivision.

buffer=nothing: heap use to do the adaptive algorithm, can be allocated using allocate_buffer, which might result in performance gain if multiple call to integrate is perform.

norm=x -> LinearAlgebra.norm(x, Inf): norm used to estimate the error.

atol=zero(T): absolute tolerance.

rtol=(atol > zero(T)) ? zero(T) : sqrt(eps(T)): relative tolerance.

maxsubdiv=8192 * 2^D: maximum number of subdivision.

HAdaptiveIntegration.map_from_reference — Method

map_from_reference(domain::DOM) where {DOM<:AbstractDomain}

Return an anonymous function that maps the reference domain to the physical domain domain.

HAdaptiveIntegration.map_to_reference — Method

map_to_reference(domain::DOM) where {DOM<:AbstractDomain}

Return an anonymous function that maps the physical domain domain to the reference domain.

Constraints:

For Orthotope, must have high_corner .> low_corner.

For Simplex{D}, the vertices must form a valid D-dimensional simplex with non-zero volume).

HAdaptiveIntegration.orders — Method

orders(rule::AR) where {AR<:AbstractRule}

Return the high and low order of the embedded cubature rule.

HAdaptiveIntegration.orthotope — Method

orthotope(low_corner, high_corner)
orthotope(T::DataType, low_corner, high_corner)

Return an axes-aligned orthotope in D dimensions, with element type T, given by two points low_corner and high_corner. Note that, we must have low_corner .≤ high_corner.

HAdaptiveIntegration.rectangle — Method

rectangle(low_corner, high_corner)
rectangle(T::DataType, low_corner, high_corner)

Return an axes-aligned rectangle, with element type T, given by two 2d-points low_corner and high_corner with. Note that, we must have low_corner .≤ high_corner.

HAdaptiveIntegration.reference_domain — Method

reference_domain(::Type{<:AbstractDomain})

Return the reference domain for the given domain type.

HAdaptiveIntegration.reference_orthotope — Method

reference_orthotope(D::Int)
reference_orthotope(T::DataType, D::Int)

Return the reference D-dimensional orthotope [0, 1]ᴰ with element type T.

HAdaptiveIntegration.reference_simplex — Method

reference_simplex(D::Int)
reference_simplex(T::DataType, D::Int)

Return the reference D-dimensional simplex with element type T, which is the convex hull of the N=D+1 points (0,...,0), (1,0,...,0), (0,1,0,...,0), ..., (0,...,0,1).

HAdaptiveIntegration.segment — Method

segment(xmin, xmax)
segment(T::DataType, xmin, xmax)

Return a segment in 1 dimensions with element type T representing the interval [xmin, xmax] with xmin ≤ xmax.

HAdaptiveIntegration.simplex — Method

simplex(vertices...)
simplex(T::DataType, vertices...)

Return a D-simplex with element type T from a collection of N=D+1 vertices. Note that all vertices must have the same length D.

HAdaptiveIntegration.subdivide_cuboid8 — Method

subdivide_cuboid8(c::Cuboid)

Divide the cuboid c into 8 cuboid by connecting the center of the cuboid to the midpoints of the edges.

HAdaptiveIntegration.subdivide_orthotope — Method

subdivide_orthotope(h::Orthotope)

Subdivide the D-orthotope h into 2ᴰ smaller orthotopes by splitting each dimension at its midpoint.

HAdaptiveIntegration.subdivide_rectangle4 — Method

subdivide_rectangle4(r::Rectangle)

Divide the rectangle r into four squares by connecting the center of the square to the midpoints of the edges.

HAdaptiveIntegration.subdivide_reference_orthotope — Method

subdivide_reference_orthotope(::Val{D}, ::Type{T}=Float64) where {D,T}

Like subdivide_orthotope, but operates on the reference orthotope. Since the output depends only on the dimension D, and the type T used to represent coordinates, this function is generated for each combination of D and T.

HAdaptiveIntegration.subdivide_reference_simplex — Method

subdivide_reference_simplex(::Val{D}, ::Type{T}=Float64) where {D,T}

Like subdivide_simplex, but operates on the reference simplex. Since the output depends only on the dimension D, and the type T used to represent coordinates, this function is generated for each combination of D and T.

HAdaptiveIntegration.subdivide_segment2 — Method

subdivide_segment2(s::Segment)

Divide the segment s into two segments of equal length.

HAdaptiveIntegration.subdivide_simplex — Method

subdivide_simplex(s::Simplex)

Subdivide a D-simplex into 2ᴰ simplices by using the Freudenthal triangulation.

Implements the RedRefinementND algorithm in Simplicial grid refinement: on Freudenthal's algorithm and the optimal number of congruence classes.

HAdaptiveIntegration.subdivide_tetrahedron8 — Method

subdivide_tetrahedron8(t::Tetrahedron)

Divide the tetrahedron t into eight tetrahedra by connecting the midpoints of the edges.

HAdaptiveIntegration.subdivide_triangle2 — Method

subdivide_triangle2(s::Triangle)

Divide the triangle t into two triangles by connecting the first point of t to the midpoints of the two other points.

HAdaptiveIntegration.subdivide_triangle4 — Method

subdivide_triangle4(t::Triangle)

Divide the triangle t into four triangles by connecting the midpoints of the edges.

HAdaptiveIntegration.tetrahedron — Method

tetrahedron(a, b, c, d)
tetrahedron(T::DataType, a, b, c, d)

Return a tetrahedron in 3 dimensions, with element type T, given by four 3d-points a, b, c, and d.

HAdaptiveIntegration.triangle — Method

triangle(a, b, c)
triangle(T::DataType, a, b, c)

Return a triangle in 2 dimensions, with element type T, given by three 2d-points a, b, and c.

HAdaptiveIntegration.validate_orders — Method

validate_orders(
    ec::EmbeddedCubature{D,T},
    ::Type{DOM},
    order_high::Int,
    order_low::Int;
    atol=zero(T),
    rtol=(atol > zero(T)) ? zero(T) : 10 * eps(T),
) where {D,T,DOM<:AbstractDomain}

Return true if the embedded cubature on the reference domain of DOM integrate exactly (within tolerance) the monomials up to degree order_high for the high order cubature and order_low for the low order cubature. Else return false.

HAdaptiveIntegration.validate_subdivision — Method

validate_subdivision(
    subdiv_algo,
    domain::AbstractDomain{D,T};
    atol=zero(T),
    rtol=(atol > zero(T)) ? zero(T) : 10 * eps(T),
) where {D,T}

Return true if the sum of the volume of the subdomain by the subdiv_algo is equal to the volume of the domain else return false.