Reference

Contents

Reference

Contents

Index

Index

HAdaptiveIntegration.Cuboid

HAdaptiveIntegration.Domain

HAdaptiveIntegration.EmbeddedCubature

HAdaptiveIntegration.Orthotope

HAdaptiveIntegration.Rectangle

HAdaptiveIntegration.Segment

HAdaptiveIntegration.Simplex

HAdaptiveIntegration.TabulatedEmbeddedCubature

HAdaptiveIntegration.Tetrahedron

HAdaptiveIntegration.Triangle

HAdaptiveIntegration._subdivide_reference_simplex

HAdaptiveIntegration.abs_det_jacobian

HAdaptiveIntegration.check_subdivision

HAdaptiveIntegration.combinations

HAdaptiveIntegration.cuboid

HAdaptiveIntegration.embedded_cubature

HAdaptiveIntegration.embedded_cubature

HAdaptiveIntegration.map_from_reference

HAdaptiveIntegration.map_to_reference

HAdaptiveIntegration.orthotope

HAdaptiveIntegration.rectangle

HAdaptiveIntegration.reference_domain

HAdaptiveIntegration.segment

HAdaptiveIntegration.simplex

HAdaptiveIntegration.subdivide_cuboid8

HAdaptiveIntegration.subdivide_rectangle4

HAdaptiveIntegration.subdivide_segment2

HAdaptiveIntegration.subdivide_segment3

HAdaptiveIntegration.subdivide_simplex

HAdaptiveIntegration.subdivide_tetrahedron8

HAdaptiveIntegration.subdivide_triangle2

HAdaptiveIntegration.subdivide_triangle4

HAdaptiveIntegration.tabulated_embedded_cubature

HAdaptiveIntegration.tetrahedron

HAdaptiveIntegration.triangle

HAdaptiveIntegration.Cuboid — Type

A axes-aligned cuboid given by two 3d-points low_corner and high_corner.

HAdaptiveIntegration.Domain — Type

abstract type Domain{D,T<:Real}

Abstract type for integration domains in D dimensions with type T<:Real.

HAdaptiveIntegration.EmbeddedCubature — Type

struct EmbeddedCubature{H,L,D,T}

An embedded cubature rule consisting of a high order cubature rule with H nodes and a low order cubature rule with L nodes. The cubature nodes and weights are assume to be for the reference simplex or orthotope. Note that the low order cubature uses nodes[1:L] as its nodes.

Fields:

nodes::SVector{H,SVector{D,T}}: the cubature nodes;

weights_high::SVector{H,T}: the cubature weights for the high order cubature;

weights_low::SVector{L,T}: the cubature weights for the low order cubature.

HAdaptiveIntegration.Orthotope — Type

struct Orthotope{D,T}

Axes-aligned Orthotope in D dimensions given by two points low_corner and high_corner.

HAdaptiveIntegration.Rectangle — Type

An axes-aligned rectangle given by two 2d-points low_corner and high_corner.

HAdaptiveIntegration.Segment — Type

A segment in 1 dimensions given by two value xmin and xmax.

HAdaptiveIntegration.Simplex — Type

struct Simplex{D,T,N}

A simplex in D dimensions with N=D+1 points of type T.

HAdaptiveIntegration.TabulatedEmbeddedCubature — Type

struct TabulatedEmbeddedCubature

An embedded cubature rule consisting of a high order cubature rule nodes and a low order cubature rule. The cubature nodes and weights are assume to be for the reference simplex or orthotope. Note that the low order cubature uses nodes[1:L] as its nodes.

Fields:

name::String: name of the embedded cubature;

domain::Domain: domain of the cubature.

reference::String: where the values are found;

nb_significant_digits::Int: number of significant digits on the node and weight values, 10^-nb_significant_digits give the relative precision of the values;

nodes::Vector{Vector{String}}: the cubature nodes;

weights_high::Vector{String}: the cubature weights for the high order cubature;

order_high::Int: order of the high order cubature;

weights_low::Vector{String}: the cubature weights for the low order cubature;

order_low::Int: order of the low order cubature.

HAdaptiveIntegration.Tetrahedron — Type

A tetrahedron in 3 dimensions with 4 points of type T.

HAdaptiveIntegration.Triangle — Type

A triangle in 2 dimensions with 3 vertices of type T.

HAdaptiveIntegration._subdivide_reference_simplex — Method

_subdivide_reference_simplex(::Val{D}, ::Type{T}=Float64)

Like subdivide_simplex, but operates on the reference simplex. Since the output depends only on the dimension D, and the type T used to represent coordinates, this function is generated for each combination of D and T.

HAdaptiveIntegration.abs_det_jacobian — Method

abs_det_jacobian(d::Domain)

The absolute value of the Jacobian's determinant of the map from the reference domain to the physical domain d.

HAdaptiveIntegration.check_subdivision — Method

check_subdivision(
    domain::Domain{D,T},
    subdiv_algo;
    atol::T=zero(T),
    rtol::T=(atol > zero(T)) ? zero(T) : 10 * eps(T),
) where {D,T<:Real}

Return nothing if the sum of the volume of the subdomain by the subdiv_algo is equal to the volume of the domain else throw an error.

HAdaptiveIntegration.combinations — Method

combinations(n, k)

Helper function to generate all combinations of k elements from 1:n, similar to calling combinations(1:n, k) from Combinatorics.jl.

HAdaptiveIntegration.cuboid — Method

cuboid(low_corner, high_corner)

A axes-aligned cuboid given by two 3d-points low_corner and high_corner.

HAdaptiveIntegration.embedded_cubature — Function

embedded_cubature(tec::TabulatedEmbeddedCubature, T=Float64)

Return the EmbeddedCubature with type T from an TabulatedEmbeddedCubature.

HAdaptiveIntegration.embedded_cubature — Method

embedded_cubature(nodes, weights_high, weights_low)

Return an embedded cubature form a vector of nodes and two vector of weights for the high order and low order cubature.

HAdaptiveIntegration.map_from_reference — Method

map_from_reference(d::Domain)::Function

Return an anonymous function that maps the reference domain to the physical domain d.

HAdaptiveIntegration.map_to_reference — Method

map_to_reference(d::Domain)::Function

Return an anonymous function that maps the physical domain d to the reference domain.

HAdaptiveIntegration.orthotope — Method

orthotope(low_corner, high_corner)

An axes-aligned orthotope in D dimensions given by two vectors low_corner and high_corner.

HAdaptiveIntegration.rectangle — Method

rectangle(low_corner, high_corner)

An axes-aligned rectangle given by two 2d-points low_corner and high_corner.

HAdaptiveIntegration.reference_domain — Method

reference_domain(domain_type::DataType)

Return the reference domain for domain_type.

Return the reference D-simplex given by the vertices (0,...,0), (1,0,...,0), (0,1,0,...,0), (0,...,0,1).

Return the reference orthotope in D dimensions, representing [0, 1]ᴰ.

HAdaptiveIntegration.segment — Method

segment(xmin, xmax)

A segment in 1 dimensions representing [xmin, xmax].

HAdaptiveIntegration.simplex — Method

simplex(points...)

A simplex in D dimensions with N=D+1 points of type T.

HAdaptiveIntegration.subdivide_cuboid8 — Method

subdivide_cuboid8(c::Cuboid)

Divide the cuboid c into 8 cuboid by connecting the center of the cuboid to the midpoints of the edges.

HAdaptiveIntegration.subdivide_rectangle4 — Method

subdivide_rectangle4(r::Rectangle)

Divide the rectangle r into four squares by connecting the center of the square to the midpoints of the edges.

HAdaptiveIntegration.subdivide_segment2 — Method

subdivide_segment2(s::Segment)

Divide the segment s into two segments of equal length.

HAdaptiveIntegration.subdivide_segment3 — Method

subdivide_segment3(s::Segment)

Divide the segment s into three segments of equal length.

HAdaptiveIntegration.subdivide_simplex — Method

subdivide_simplex(s::Simplex)

Subdivive a D-dimensional simplex into 2ᴰ simplices by using the Freudenthal triangulation.

Implements the RedRefinementND` algorithm in Simplicial grid refinement: on Freudenthal's algorithm and the optimal number of congruence classes.

HAdaptiveIntegration.subdivide_tetrahedron8 — Method

subdivide_tetrahedron8(t::Tetrahedron)

Divide the tetrahedron t into eight tetrahedra by connecting the midpoints of the edges.

HAdaptiveIntegration.subdivide_triangle2 — Method

subdivide_triangle2(s::Triangle)

Divide the triangle t into two triangles by connecting the first point of t to the midpoints of the two other points.

HAdaptiveIntegration.subdivide_triangle4 — Method

subdivide_triangle4(t::Triangle)

Divide the triangle t into four triangles by connecting the midpoints of the edges.

HAdaptiveIntegration.tabulated_embedded_cubature — Method

tabulated_embedded_cubature(;
    name::String,
    domain::Domain,
    reference::String,
    nb_significant_digits::Int,
    nodes::Vector{Vector{String}},
    weights_high::Vector{String},
    order_high::Int,
    weights_low::Vector{String},
    order_low::Int,
)

An embedded cubature rule consisting of a high order cubature rule nodes and a low order cubature rule. The cubature nodes and weights are assume to be for the reference simplex or orthotope. Note that the low order cubature uses nodes[1:L] as its nodes.

Fields:

name::String: name of the embedded cubature;

domain::Domain: domain of the cubature.

reference::String: where the values are found;

nb_significant_digits::Int: number of significant digits on the node and weight values, 10^-nb_significant_digits give the relative precision of the values;

nodes::Vector{Vector{String}}: the cubature nodes;

weights_high::Vector{String}: the cubature weights for the high order cubature;

order_high::Int: order of the high order cubature;

weights_low::Vector{String}: the cubature weights for the low order cubature;

order_low::Int: order of the low order cubature.

HAdaptiveIntegration.tetrahedron — Method

tetrahedron(a, b, c, d)

A tetrahedron in 3 dimensions given by the 3d-points a, b, c, and d.

HAdaptiveIntegration.triangle — Method

triangle(a, b, c)

A triangle in 2 dimensions given by the 2d-points a, b, and c.