Docstrings

HAdaptiveIntegration.CUBE_BE65 — Constant

Berntsen and Espelid with 65 nodes.

HAdaptiveIntegration.SEGMENT_GK15 — Constant

Gauss-Kronrod with 15 nodes.

HAdaptiveIntegration.SEGMENT_GK31 — Constant

Gauss-Kronrod with 31 nodes.

HAdaptiveIntegration.SQUARE_CHG25 — Constant

Cools and Haegemans with 25 nodes.

HAdaptiveIntegration.TETRAHEDRON_GM35 — Constant

Grundmann and Möller with 35 nodes.

HAdaptiveIntegration.TRIANGLE_GM20 — Constant

Grundmann and Möller with 20 nodes.

HAdaptiveIntegration.TRIANGLE_RL19 — Constant

Laurie Radon with 19 nodes.

HAdaptiveIntegration.Cuboid — Type

Cuboid{T} = Orthotope{3,T}

HAdaptiveIntegration.Domain — Type

abstract type Domain{D,T<:Real}

Abstract type for integration on domains in D dimensions with value type T.

HAdaptiveIntegration.EmbeddedCubature — Type

struct EmbeddedCubature{H,L,D,T}

An embedded cubature rule consisting of a high order cubature rule with H nodes and a low order cubature rule with L nodes. Note that the low order cubature uses nodes[1:L] as its nodes. The cubature nodes and weights are assume to be for the reference domain.

Fields:

nodes::SVector{H,SVector{D,T}}: the cubature nodes.

weights_high::SVector{H,T}: the cubature weights for the high order cubature.

weights_low::SVector{L,T}: the cubature weights for the low order cubature.

HAdaptiveIntegration.EmbeddedCubature — Method

(ec::EmbeddedCubature{H,L,D,T})(
    fct, domain::Domain{D,T}, norm=x -> LinearAlgebra.norm(x, Inf)
) where {H,L,D,T}

Return I_high and norm(I_high - I_low) where I_high and I_low are the result of the high order cubature and the low order cubature on domain. The function fct must take a SVector{D,T} to a return type K, and K must support the multiplication by a scalar and the addition. Note that there is no check, beyond compatibility of dimension and type, that the embedded cubature is for the right domain.

HAdaptiveIntegration.GrundmannMoeller — Type

struct GrundmannMoeller

Cubature rule for a dim-simplex of degree deg.

HAdaptiveIntegration.Orthotope — Type

struct Orthotope{D,T} <: Domain{D,T}

Axes-aligned Orthotope in D dimensions, with value type T, given by two points low_corner and high_corner. Note that, we must have low_corner .≤ high_corner.

Fields:

low_corner::SVector{D,T}: the low corner.

high_corner::SVector{D,T}: the high corner.

HAdaptiveIntegration.Rectangle — Type

Rectangle{T} = Orthotope{2,T}

HAdaptiveIntegration.Segment — Type

Segment{T} = Orthotope{1,T}

HAdaptiveIntegration.Simplex — Type

struct Simplex{D,T,N} <: Domain{D,T}

A simplex in D dimensions with N=D+1 vertices of value type T.

Fields:

vertices::SVector{N,SVector{D,T}}: vertices of the simplex.

HAdaptiveIntegration.TabulatedEmbeddedCubature — Type

struct TabulatedEmbeddedCubature

An embedded cubature rule consisting of a high order cubature rule and a low order cubature rule. Note that the low order cubature uses nodes[1:L] as its nodes where L is the length of the weights_low. The cubature nodes and weights are assume to be for the reference domain.

Fields:

description::String: description of the embedded cubature.

domain::String: domain of the cubature, the possible values are "reference segment", "reference rectangle", "reference cuboid", "reference triangle", and "reference tetrahedron".

reference::String: where the values are found.

nb_significant_digits::Int: number of significant digits on the node and weight values, 10^-nb_significant_digits give the relative precision of the values.

nodes::Vector{Vector{String}}: the cubature nodes.

weights_high::Vector{String}: the cubature weights for the high order cubature.

order_high::Int: order of the high order cubature.

weights_low::Vector{String}: the cubature weights for the low order cubature.

order_low::Int: order of the low order cubature.

HAdaptiveIntegration.Tetrahedron — Type

Tetrahedron{T} = Simplex{3,T,4}

HAdaptiveIntegration.Triangle — Type

Triangle{T} = Simplex{2,T,3}

HAdaptiveIntegration.abs_det_jac — Method

abs_det_jac(domain::Domain)

Return the absolute value of the Jacobian's determinant of the map from the reference domain to the physical domain domain.

HAdaptiveIntegration.allocate_buffer — Method

allocate_buffer(
    fct, domain::DOM, ec::EmbeddedCubature=default_embedded_cubature(domain)
) where {DOM<:Domain}

Return a buffer which can be pass to the integrate function for improve allocation if called multiple times.

HAdaptiveIntegration.check_order — Method

check_order(
    tec::TabulatedEmbeddedCubature,
    domain::Domain{D,T};
    atol=zero(T),
    rtol=(atol > zero(T)) ? zero(T) : 10 * eps(T),
) where {D,T}

Return 0 if the tabulated embedded cubature on the reference domain of domain integrate exactly (within tolerance) the monomials up to degree order_high for the high oder cubature and order_low for the low order cubature. Else return 1.

HAdaptiveIntegration.check_order — Method

check_order(
    ec::EmbeddedCubature{H,L,D,T},
    domain::Domain{D,T},
    order_high::Int,
    order_low::Int;
    atol=zero(T),
    rtol=(atol > zero(T)) ? zero(T) : 10 * eps(T),
) where {H,L,D,T}

Return 0 if the embedded cubature on the reference domain of domain integrate exactly (within tolerance) the monomials up to degree order_high for the high oder cubature and order_low for the low order cubature. Else return 1.

HAdaptiveIntegration.check_subdivision — Method

check_subdivision(
    subdiv_algo,
    domain::Domain{D,T};
    atol=zero(T),
    rtol=(atol > zero(T)) ? zero(T) : 10 * eps(T),
) where {D,T}

Return 0 if the sum of the volume of the subdomain by the subdiv_algo is equal to the volume of the domain else return 1.

HAdaptiveIntegration.combinations — Method

combinations(n::Int, k::Int)

Helper function to generate all combinations of k elements from 1:n, similar to calling combinations(1:n, k) from Combinatorics.jl.

HAdaptiveIntegration.cuboid — Method

cuboid([T=Float64,] low_corner, high_corner)

Return an axes-aligned cuboid, with value type T, given by two 3d-points low_corner and high_corner. Note that, we must have low_corner .≤ high_corner.

HAdaptiveIntegration.default_embedded_cubature — Method

default_embedded_cubature(domain::Domain)

Return a default embedded cubature for the domains:

dimension 1:

segment: SEGMENT_GK15

dimension 2:

rectangle: SQUARE_CHG25

triangle: TRIANGLE_RL19

dimension 3:

cuboid: CUBE_BE65

tetrahedron: TETRAHEDRON_GM35

dimension d:

simplex: [GrundmannMoeller](@ref)(d, 7)

HAdaptiveIntegration.default_subdivision — Method

default_subdivision(domain::Domain)

Return the default algorithm to subdivide domain.

dimension 1:

segment: subdivide_segment2

dimension 2:

rectangle: subdivide_rectangle4

triangle: subdivide_triangle4

dimension 3:

cuboid: subdivide_cuboid8

tetrahedron: subdivide_tetrahedron8

dimension d:

simplex: subdivide_simplex

HAdaptiveIntegration.embedded_cubature — Function

embedded_cubature(tec::TabulatedEmbeddedCubature, T::DataType=Float64)

Return the embedded cubature with value type T from a TabulatedEmbeddedCubature.

HAdaptiveIntegration.embedded_cubature — Function

embedded_cubature(gm::GrundmannMoeller, T::DataType=Float64)

Return the embedded cubature of Grundmann and Möller.

HAdaptiveIntegration.embedded_cubature — Method

embedded_cubature(
    nodes::Vector{VT}, weights_high::VT, weights_low::VT
) where {VT<:AbstractVector{<:Real}}

Return an embedded cubature form a vector of nodes and two vector of weights for the high order and low order cubature.

HAdaptiveIntegration.integral_monomial — Method

integral_monomial(domain::Domain{D,T}, tot_deg_max::Int) where {D,T}

Return the values of monomial's integral over the reference domain. It return a Vector{ Vector{ Pair{ Ntuple{dim,Int}, Rational{Int} } } }. The outer vector is index by the total degree + 1, for the total degree form 0 to tot_deg_max. The inner vector contain Pair{ Ntuple{dim,Int}, Rational{Int} } where the Ntuple{D,Int} is the multi-index of the monomial and Rational{Int} is the value of the integral.

HAdaptiveIntegration.integrate — Method

integrate(
    fct,
    domain::Domain{D,T};
    embedded_cubature::EmbeddedCubature{H,L,D,T}=default_embedded_cubature(domain),
    subdiv_algo=default_subdivision(domain),
    buffer=nothing,
    norm=x -> LinearAlgebra.norm(x, Inf),
    atol=zero(T),
    rtol=(atol > zero(T)) ? zero(T) : sqrt(eps(T)),
    maxsubdiv=512 * 2^D,
) where {H,L,D,T}

Return I and E where I is the integral of the function fct over domain and E is an error estimate.

Arguments

fct: a function that must take a SVector{D,T} to a return type K, with K must support the multiplication by a scalar of type T and the addition.

domain::Domain{D,T}: the integration domain. Currently, we support segment, triangle, rectangle, tetrahedron, cuboid, and d-dimensional simplex.

Optional arguments

ec::EmbeddedCubature{H,L,D,T}=default_embedded_cubature(domain): the embedded cubature, each supported domain has a default_embedded_cubature.

subdiv_algo=default_subdivision(domain): the subdivision algorithm, each domain has a default_subdivision.

buffer=nothing: heap use to do the adaptive algorithm, can be allocated using allocate_buffer, which might result in performance gain if multiple call to integrate is perform.

norm=x -> LinearAlgebra.norm(x, Inf): norm used to estimate the error.

atol=zero(T): absolute tolerance.

rtol=(atol > zero(T)) ? zero(T) : sqrt(eps(T)): relative tolerance.

maxsubdiv=512 * 2^D: maximum number of subdivision.

HAdaptiveIntegration.map_from_reference — Method

map_from_reference(domain::Domain)

Return an anonymous function that maps the reference domain to the physical domain domain.

HAdaptiveIntegration.map_to_reference — Method

map_to_reference(domain::Domain)

Return an anonymous function that maps the physical domain domain to the reference domain.

HAdaptiveIntegration.orthotope — Method

orthotope([T=Float64,] low_corner, high_corner)

Return an axes-aligned orthotope in D dimensions, with value type T, given by two points low_corner and high_corner with value type T. Note that, we must have low_corner .≤ high_corner.

HAdaptiveIntegration.rectangle — Method

rectangle([T=Float64,] low_corner, high_corner)

Return an axes-aligned rectangle, with value type T, given by two 2d-points low_corner and high_corner with. Note that, we must have low_corner .≤ high_corner.

HAdaptiveIntegration.reference_orthotope — Method

reference_orthotope([T=Float64,] D::Int)

Return the reference D-dimensional orthotope [0, 1]ᴰ with value type T.

HAdaptiveIntegration.reference_simplex — Method

reference_simplex([T=Float64,] D::Int)

Return the reference D-dimensional simplex with value type T, which is the convex hull of the N=D+1 points (0,...,0), (1,0,...,0), (0,1,0,...,0), ..., (0,...,0,1).

HAdaptiveIntegration.segment — Method

segment([T=Float64,] xmin, xmax)

Return a segment in 1 dimensions with value type T representing the interval [xmin, xmax] with xmin ≤ xmax.

HAdaptiveIntegration.simplex — Method

simplex([T=Float64,] vertices...)

Return a D-simplex with value type T from a collection of vertices. Note that all points must have the same length D and there must be N=D+1 points.

HAdaptiveIntegration.subdivide_cuboid8 — Method

subdivide_cuboid8(c::Cuboid)

Divide the cuboid c into 8 cuboid by connecting the center of the cuboid to the midpoints of the edges.

HAdaptiveIntegration.subdivide_rectangle4 — Method

subdivide_rectangle4(r::Rectangle)

Divide the rectangle r into four squares by connecting the center of the square to the midpoints of the edges.

HAdaptiveIntegration.subdivide_reference_simplex — Method

subdivide_reference_simplex(::Val{D}, ::Type{T}=Float64) where {D,T}

Like subdivide_simplex, but operates on the reference simplex. Since the output depends only on the dimension D, and the type T used to represent coordinates, this function is generated for each combination of D and T.

HAdaptiveIntegration.subdivide_segment2 — Method

subdivide_segment2(s::Segment)

Divide the segment s into two segments of equal length.

HAdaptiveIntegration.subdivide_simplex — Method

subdivide_simplex(s::Simplex)

Subdivide a D-simplex into 2ᴰ simplices by using the Freudenthal triangulation.

Implements the RedRefinementND algorithm in Simplicial grid refinement: on Freudenthal's algorithm and the optimal number of congruence classes.

HAdaptiveIntegration.subdivide_tetrahedron8 — Method

subdivide_tetrahedron8(t::Tetrahedron)

Divide the tetrahedron t into eight tetrahedra by connecting the midpoints of the edges.

HAdaptiveIntegration.subdivide_triangle2 — Method

subdivide_triangle2(s::Triangle)

Divide the triangle t into two triangles by connecting the first point of t to the midpoints of the two other points.

HAdaptiveIntegration.subdivide_triangle4 — Method

subdivide_triangle4(t::Triangle)

Divide the triangle t into four triangles by connecting the midpoints of the edges.

HAdaptiveIntegration.tetrahedron — Method

tetrahedron([T=Float64,] a, b, c, d)

Return a tetrahedron in 3 dimensions, with value type T, given by four 3d-points a, b, c, and d.

HAdaptiveIntegration.triangle — Method

triangle([T=Float64,] a, b, c)

Return a triangle in 2 dimensions, with value type T, given by three 2d-points a, b, and c.